+++ backend/api/src/main/java/br/edu/ifce/api/controllers/AuthenticatedUserController.java
@@ -9,13 +9,13 @@ import br.edu.ifce.usecase.ports.driver.ListTheMedicalCareUnitsInTheCity;
 import br.edu.ifce.usecase.ports.driver.RefreshUserAuthToken;
 import br.edu.ifce.usecase.ports.driver.UpdateAuthenticatedUserAddress;
 import br.edu.ifce.usecase.ports.driver.UpdateAuthenticatedUserProfile;
-import br.edu.ifce.usecase.ports.dto.MedicalCareUnityInfo;
-import br.edu.ifce.usecase.ports.dto.VaccinationData;
-import br.edu.ifce.usecase.ports.dto.addressdtos.FullAddressDTO;
-import br.edu.ifce.usecase.ports.dto.addressdtos.UpdateAddressDTO;
-import br.edu.ifce.usecase.ports.dto.messagedtos.NewMessageDTO;
-import br.edu.ifce.usecase.ports.dto.userdtos.UpdateUserDTO;
-import br.edu.ifce.usecase.ports.dto.userdtos.UserDTO;
+import br.edu.ifce.usecase.ports.responses.MedicalCareUnityInfo;
+import br.edu.ifce.usecase.ports.responses.VaccinationData;
+import br.edu.ifce.usecase.ports.responses.FullAddressDTO;
+import br.edu.ifce.usecase.ports.requests.UpdateAddressDTO;
+import br.edu.ifce.usecase.ports.requests.NewMessageDTO;
+import br.edu.ifce.usecase.ports.requests.UpdateUserDTO;
+import br.edu.ifce.usecase.ports.responses.UserDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
