+++ backend/api/src/main/java/br/edu/ifce/api/controllers/MessageManagementController.java
@@ -3,8 +3,8 @@ package br.edu.ifce.api.controllers;
 import br.edu.ifce.api.docs.MessageManagementControllerDocs;
 import br.edu.ifce.usecase.ports.driver.GetAMessageById;
 import br.edu.ifce.usecase.ports.driver.GetAllMessages;
-import br.edu.ifce.usecase.ports.dto.messagedtos.MessageDTO;
-import br.edu.ifce.usecase.ports.dto.messagedtos.MessageWithContentDTO;
+import br.edu.ifce.usecase.ports.responses.MessageDTO;
+import br.edu.ifce.usecase.ports.responses.MessageWithContentDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
