+++ backend/api/src/main/java/br/edu/ifce/api/controllers/PostController.java
@@ -3,8 +3,8 @@ package br.edu.ifce.api.controllers;
 import br.edu.ifce.api.docs.PostControllerDocs;
 import br.edu.ifce.usecase.ports.driver.GetAPostById;
 import br.edu.ifce.usecase.ports.driver.GetAllPosts;
-import br.edu.ifce.usecase.ports.dto.postdtos.PostDTO;
-import br.edu.ifce.usecase.ports.dto.postdtos.PostWithContentDTO;
+import br.edu.ifce.usecase.ports.responses.PostDTO;
+import br.edu.ifce.usecase.ports.responses.PostWithContentDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
