+++ backend/api/src/main/java/br/edu/ifce/api/controllers/UserController.java
@@ -6,9 +6,9 @@ import br.edu.ifce.usecase.ports.driver.RegisterAUser;
 import br.edu.ifce.usecase.ports.driver.ResendAccountConfirmationEmail;
 import br.edu.ifce.usecase.ports.driver.ResetUserPassword;
 import br.edu.ifce.usecase.ports.driver.SendLinkToChangePassword;
-import br.edu.ifce.usecase.ports.dto.userdtos.EmailDTO;
-import br.edu.ifce.usecase.ports.dto.userdtos.PasswordDTO;
-import br.edu.ifce.usecase.ports.dto.userdtos.UserRegistrationDTO;
+import br.edu.ifce.usecase.ports.requests.EmailDTO;
+import br.edu.ifce.usecase.ports.requests.PasswordDTO;
+import br.edu.ifce.usecase.ports.requests.UserRegistrationDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
