+++ backend/api/src/main/java/br/edu/ifce/api/docs/UserControllerDocs.java
@@ -1,8 +1,8 @@
 package br.edu.ifce.api.docs;
 
-import br.edu.ifce.usecase.ports.dto.userdtos.EmailDTO;
-import br.edu.ifce.usecase.ports.dto.userdtos.PasswordDTO;
-import br.edu.ifce.usecase.ports.dto.userdtos.UserRegistrationDTO;
+import br.edu.ifce.usecase.ports.requests.EmailDTO;
+import br.edu.ifce.usecase.ports.requests.PasswordDTO;
+import br.edu.ifce.usecase.ports.requests.UserRegistrationDTO;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
