+++ backend/usecase/src/main/java/br/edu/ifce/usecase/ports/driver/GetDataFromTheVaccinationCampaign.java
@@ -1,6 +1,6 @@
 package br.edu.ifce.usecase.ports.driver;
 
-import br.edu.ifce.usecase.ports.dto.VaccinationData;
+import br.edu.ifce.usecase.ports.responses.VaccinationData;
 
 public interface GetDataFromTheVaccinationCampaign {
     VaccinationData execute();
