+++ backend/usecase/src/main/java/br/edu/ifce/usecase/ports/driver/GetInformationAboutZipCode.java
@@ -1,6 +1,6 @@
 package br.edu.ifce.usecase.ports.driver;
 
-import br.edu.ifce.usecase.ports.dto.CompleteZipCodeInformation;
+import br.edu.ifce.usecase.ports.responses.CompleteZipCodeInformation;
 
 public interface GetInformationAboutZipCode {
     CompleteZipCodeInformation execute(String zip);
