+++ backend/usecase/src/main/java/br/edu/ifce/usecase/ports/driver/ListTheMedicalCareUnitsInTheCity.java
@@ -1,6 +1,6 @@
 package br.edu.ifce.usecase.ports.driver;
 
-import br.edu.ifce.usecase.ports.dto.MedicalCareUnityInfo;
+import br.edu.ifce.usecase.ports.responses.MedicalCareUnityInfo;
 
 public interface ListTheMedicalCareUnitsInTheCity {
     MedicalCareUnityInfo execute();
