+++ backend/usecase/src/main/java/br/edu/ifce/usecase/ports/requests/NewMessageDTO.java
@@ -1,4 +1,4 @@
-package br.edu.ifce.usecase.ports.dto.messagedtos;
+package br.edu.ifce.usecase.ports.requests;
 
 import br.edu.ifce.domain.Message;
 import lombok.Getter;
