+++ backend/usecase/src/main/java/br/edu/ifce/usecase/ports/requests/UpdateUserDTO.java
@@ -1,4 +1,4 @@
-package br.edu.ifce.usecase.ports.dto.userdtos;
+package br.edu.ifce.usecase.ports.requests;
 
 import br.edu.ifce.domain.User;
 import lombok.Getter;
