+++ backend/usecase/src/main/java/br/edu/ifce/usecase/ports/responses/PostDTO.java
@@ -1,4 +1,4 @@
-package br.edu.ifce.usecase.ports.dto.postdtos;
+package br.edu.ifce.usecase.ports.responses;
 
 import br.edu.ifce.domain.Post;
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -9,21 +9,19 @@ import java.time.LocalDateTime;
 
 @Getter
 @Setter
-public class PostWithContentDTO {
+public class PostDTO {
 
     private Long id;
     private String title;
     private String description;
-    private String content;
 
     @JsonFormat(pattern = "dd/MM/yyyy HH:mm")
     private LocalDateTime createdAt;
 
-    public PostWithContentDTO (Post post) {
+    public PostDTO(Post post) {
         this.id = post.getId();
         this.title = post.getTitle();
         this.description = post.getDescription();
-        this.content = post.getContent();
         this.createdAt = post.getCreatedAt();
     }
 }
